Don't always rebuild external modules if kernel image not rebuilt
The code in kernel-gen includes a timestamp-based check to avoid invoking kmake in case .config hasn't changed.
This can be fooled if .config is modified in a way that doesn't cause kmake to rebuild the kernel image.
Avoid this by always touching the kernel image after a kmake that we determined was necessary.